backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# ML/AI
open-clip-torch==2.23.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3

# Dimensionality reduction and kNN
umap-learn==0.5.5
faiss-cpu==1.7.4
scikit-learn==1.3.2

# AWS
boto3==1.34.10
botocore==1.34.10

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
opencv-python-headless==4.8.1.78

